+++ src/ansys/mechanical/core/run.py
@@ -171,6 +171,8 @@ def cli(
     if input_script:
         args.append("-script")
         args.append(input_script)
+
+    if (not graphical) and input_script and (version < 241):
         warnings.warn(
             "Please ensure ExtAPI.Application.Close() is at the end of your script. "
             "Without this command, Batch mode will not terminate.",