+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG](https://kee
 ### Fixed
 
 -   Broken links (#316)
+-   Remove project lock file on close (#320)
 
 ## [0.9.3](https://github.com/ansys/pymechanical/releases/tag/v0.9.3) - July 27 2023