+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG](https://kee
 
 -   Added warning for ansys-mechanical when provided an input script (#319)
 -   Add changelog check to CI/CD (#322)
+-   Added version check for ansys-mechanical warning message (#323)
 
 ### Changed