+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This document follows the conventions laid out in [Keep a CHANGELOG](https://kee
 
 ### Added
 
+-   Added warning for ansys-mechanical when provided an input script (#319)
+-   Add changelog check to CI/CD (#322)
+-   Added version check for ansys-mechanical warning message (#323)
+
 ### Changed
 
 ### Fixed