+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG](https://kee
 
 -   Broken links (#316)
 -   Remove project lock file on close (#320)
+-   Fixed warning message for ansys-mechanical (#326)
 
 ## [0.9.3](https://github.com/ansys/pymechanical/releases/tag/v0.9.3) - July 27 2023