+++ CHANGELOG.md
@@ -16,6 +16,8 @@ This document follows the conventions laid out in [Keep a CHANGELOG](https://kee
 
 ### Changed
 
+-   Update python minimum requirement from 3.7 to 3.8 (#333)
+
 ### Fixed
 
 -   Broken links (#316)